初回再生時のプレイリスト生成責務分離
